Set Integer as id in JPA repos

+++ src/main/java/pl/miloszlewandowski/backtrackedpromisesparadaisehotel/model/GuestRepository.java
@@ -2,6 +2,6 @@ package pl.miloszlewandowski.backtrackedpromisesparadaisehotel.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GuestRepository extends JpaRepository<Guest, Long> {
+public interface GuestRepository extends JpaRepository<Guest, Integer> {
 
 }
