+++ src/main/java/pl/miloszlewandowski/backtrackedpromisesparadaisehotel/model/RoomRepository.java
@@ -2,6 +2,6 @@ package pl.miloszlewandowski.backtrackedpromisesparadaisehotel.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Guest, Long> {
+public interface RoomRepository extends JpaRepository<Guest, Integer> {
 
 }
